+++ packages/terra-date-time-picker/tests/jest/DateTimePicker.test.jsx
@@ -79,11 +79,11 @@ it('Should not throw any errors while date value is outside of the Min, Max rang
     <DateTimePicker initialTimeZone="America/Chicago" name="date-time-input" maxDate="2017-04-01T10:30" minDate="2017-04-10T10:30" />
   ));
 
-  const dateMonthInput = datePicker.find({ name: 'terra-date-month-input', type: 'text' }).at(0);
+  const dateMonthInput = datePicker.find({ name: 'terra-date-month-input', type: 'number' }).at(0);
   dateMonthInput.simulate('change', { target: { value: '04' } });
-  const dateDayInput = datePicker.find({ name: 'terra-date-day-input', type: 'text' }).at(0);
+  const dateDayInput = datePicker.find({ name: 'terra-date-day-input', type: 'number' }).at(0);
   dateDayInput.simulate('change', { target: { value: '12' } });
-  const dateYearInput = datePicker.find({ name: 'terra-date-year-input', type: 'text' }).at(0);
+  const dateYearInput = datePicker.find({ name: 'terra-date-year-input', type: 'number' }).at(0);
   dateYearInput.simulate('change', { target: { value: '2017' } });
   expect(() => {
     const hourInput = datePicker.find({ name: 'terra-time-hour-input' }).at(2);
@@ -149,6 +149,15 @@ it('should validate the createSafeDate helper', () => {
   expect(DateTimeUtils.createSafeDate('2019-06-10T16:00:01', 'America/Test')).not.toEqual(momentDate);
   // Invalid date and invalid timezone
   expect(DateTimeUtils.createSafeDate('2019-06-35T16:00:01', 'America/Test')).toEqual(undefined);
+});
+
+it('should render a default date time picker with ariaLabel', () => {
+  const ariaLabel = 'Select date and time';
+  const datePicker = enzymeIntl.shallowWithIntl(
+    <DateTimePicker initialTimeZone="America/Chicago" name="date-time-input" ariaLabel={ariaLabel} />,
+  );
+  expect(datePicker.prop('ariaLabel')).toBe(ariaLabel);
+  expect(datePicker).toMatchSnapshot();
 });
 
 it('correctly applies the theme context className', () => {