+++ packages/terra-date-time-picker/tests/wdio/date-time-picker-spec.js
@@ -387,19 +387,14 @@ Terra.describeViewports('DateTimePicker', ['tiny', 'large'], () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/date-time-picker/date-time-picker-on-change-raw');
       Terra.hideInputCaret('input[name="terra-date-year-input"]');
       Terra.hideInputCaret('input[name="terra-time-minute-input"]');
-      $('input[name="terra-date-month-input"]').setValue('07');
-      $('input[name="terra-date-day-input"]').setValue('12');
+      $('input[name="terra-date-day-input"]').setValue('01');
       $('input[name="terra-time-hour-input"]').setValue('10');
       $('input[name="terra-time-minute-input"]').setValue('30');
-      $('input[name="terra-date-month-input"]').click();
-      browser.keys('Delete');
-      $('input[name="terra-date-day-input"]').click();
-      browser.keys('Delete');
       $('input[name="terra-date-year-input"]').setValue('2017');
 
       expect($('#iso').getText()).toEqual('');
-      expect($('#input-value').getText()).toEqual('//2017 10:30');
-      expect($('#date-value').getText()).toEqual('//2017');
+      expect($('#input-value').getText()).toEqual('/01/2017 10:30');
+      expect($('#date-value').getText()).toEqual('/01/2017');
       expect($('#time-value').getText()).toEqual('10:30');
       expect($('#ambiguous-date').getText()).toEqual('No');
       expect($('#complete-date').getText()).toEqual('No');
@@ -695,6 +690,10 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
           expect($('input[name="terra-time-hour-input"]')).toHaveValue((`0${today.hour()}`).slice(-2));
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
+          const minuteValue = Number($('input[name="terra-time-minute-input"]').getValue());
+          const minuteTimeValue = today.minutes();
+          const minuteInRange = (minuteTimeValue === minuteValue + 1 || minuteTimeValue === minuteValue);
+          expect(minuteInRange).toBe(true);
         });
 
         it('should set date to today in the second input', () => {
@@ -711,7 +710,7 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
           const secondsValue = Number($('input[name="terra-time-second-input"]').getValue());
           const secondsTimeValue = today.seconds();
-          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue);
+          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue + 2 || secondsTimeValue === secondsValue);
           expect(secondsInRange).toBe(true);
         });
       });
@@ -771,7 +770,7 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
           const secondsValue = Number($('input[name="terra-time-second-input"]').getValue());
           const secondsTimeValue = today.seconds();
-          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue);
+          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue + 2 || secondsTimeValue === secondsValue);
           expect(secondsInRange).toBe(true);
         });
       });
@@ -819,7 +818,7 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
           const secondsValue = Number($('input[name="terra-time-second-input"]').getValue());
           const secondsTimeValue = today.seconds();
-          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue);
+          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue + 2 || secondsTimeValue === secondsValue);
           expect(secondsInRange).toBe(true);
         });
       });
@@ -827,7 +826,7 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
 
     describe('-', () => {
       describe('if blank date', () => {
-        it('should set date to today and now minus 1 minute in the hour input', () => {
+        it('should set date to today and now minus 1 hour in the hour input', () => {
           browser.url('/raw/tests/cerner-terra-framework-docs/date-time-picker/date-time-picker-default');
           browser.refresh();
           $('input[name="terra-time-hour-input"]').click();
@@ -841,7 +840,7 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
           const hourValue = Number($('input[name="terra-time-hour-input"]').getValue());
           const hourTimeValue = today.hours();
-          const hourInRange = (hourTimeValue === hourValue + 1 || hourTimeValue === hourValue);
+          const hourInRange = (hourTimeValue === hourValue - 1 || hourTimeValue === hourValue);
           expect(hourInRange).toBe(true);
         });
 
@@ -859,17 +858,17 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
           const minuteValue = Number($('input[name="terra-time-minute-input"]').getValue());
           const minuteTimeValue = today.minutes();
-          const minuteInRange = (minuteTimeValue === minuteValue + 1 || minuteTimeValue === minuteValue);
+          const minuteInRange = (minuteTimeValue === minuteValue - 1 || minuteTimeValue === minuteValue);
           expect(minuteInRange).toBe(true);
         });
 
-        it('should set date to today and now minus 1 minute in the second input', () => {
+        it('should set date to today and now minus 1 second in the second input', () => {
           browser.url('/raw/tests/cerner-terra-framework-docs/date-time-picker/date-time-picker-with-seconds');
           browser.refresh();
           $('input[name="terra-time-second-input"]').click();
           browser.keys('-');
 
-          const today = moment.utc().subtract(1, 'minutes');
+          const today = moment.utc().subtract(1, 'seconds');
           expect($('input[name="terra-date-year-input"]')).toHaveValue(today.year().toString());
           expect($('input[name="terra-date-month-input"]')).toHaveValue((`0${(today.month() + 1)}`).slice(-2));
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
@@ -915,7 +914,7 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
           const minuteValue = Number($('input[name="terra-time-minute-input"]').getValue());
           const minuteTimeValue = today.minutes();
-          const minuteInRange = (minuteTimeValue === minuteValue + 1 || minuteTimeValue === minuteValue);
+          const minuteInRange = (minuteTimeValue === minuteValue - 1 || minuteTimeValue === minuteValue);
           expect(minuteInRange).toBe(true);
         });
 
@@ -1150,7 +1149,7 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
 
     describe('+', () => {
       describe('if blank date', () => {
-        it('should set date to today and now plus 1 minute in the hour input', () => {
+        it('should set date to today and now plus 1 hour in the hour input', () => {
           browser.url('/raw/tests/cerner-terra-framework-docs/date-time-picker/date-time-picker-default');
           browser.refresh();
           $('input[name="terra-time-hour-input"]').click();
@@ -1186,21 +1185,22 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect(minuteInRange).toBe(true);
         });
 
-        it('should set date to today and now plus 1 minute in the second input', () => {
+        it('should set date to today and now plus 1 seconds in the second input', () => {
           browser.url('/raw/tests/cerner-terra-framework-docs/date-time-picker/date-time-picker-with-seconds');
           browser.refresh();
           $('input[name="terra-time-second-input"]').click();
           browser.keys('+');
 
-          const today = moment.utc().add(1, 'minutes');
+          const today = moment.utc().add(1, 'seconds');
           expect($('input[name="terra-date-year-input"]')).toHaveValue(today.year().toString());
           expect($('input[name="terra-date-month-input"]')).toHaveValue((`0${(today.month() + 1)}`).slice(-2));
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
           expect($('input[name="terra-time-hour-input"]')).toHaveValue((`0${today.hour()}`).slice(-2));
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
+          expect($('input[name="terra-time-second-input"]')).toHaveValue((`0${today.second()}`).slice(-2));
           const secondsValue = Number($('input[name="terra-time-second-input"]').getValue());
           const secondsTimeValue = today.seconds();
-          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue);
+          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue + 2 || secondsTimeValue === secondsValue);
           expect(secondsInRange).toBe(true);
         });
       });
@@ -1246,7 +1246,7 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect(minuteInRange).toBe(true);
         });
 
-        it('should set date to today and now plus 1 minute in the second input', () => {
+        it('should set date to today and now plus 1 second in the second input', () => {
           browser.url('/raw/tests/cerner-terra-framework-docs/date-time-picker/date-time-picker-with-seconds');
           browser.refresh();
           $('input[name="terra-date-year-input"]').click();
@@ -1256,15 +1256,16 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           $('input[name="terra-time-second-input"]').click();
           browser.keys('+');
 
-          const today = moment.utc().add(1, 'minutes');
+          const today = moment.utc().add(1, 'seconds');
           expect($('input[name="terra-date-year-input"]')).toHaveValue(today.year().toString());
           expect($('input[name="terra-date-month-input"]')).toHaveValue((`0${(today.month() + 1)}`).slice(-2));
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
           expect($('input[name="terra-time-hour-input"]')).toHaveValue((`0${today.hour()}`).slice(-2));
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
+          expect($('input[name="terra-time-second-input"]')).toHaveValue((`0${today.second()}`).slice(-2));
           const secondsValue = Number($('input[name="terra-time-second-input"]').getValue());
           const secondsTimeValue = today.seconds();
-          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue);
+          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue + 2 || secondsTimeValue === secondsValue);
           expect(secondsInRange).toBe(true);
         });
       });