+++ packages/terra-framework-docs/src/terra-dev-site/doc/date-time-picker/example/DateTimePickerTimeVariant.jsx
@@ -19,6 +19,7 @@ const DateTimePickerExampleTimeVariant = () => {
       </p>
       <Field label="Enter Date/Time" htmlFor="timeVariant">
         <DateTimePicker
+          ariaLabel="Enter Date/Time"
           name="date-time-picker-example"
           dateInputAttributes={{ id: 'timeVariant' }}
           onChange={handleDateTimeChange}