+++ packages/terra-modal-manager/src/ModalManager.jsx
@@ -30,6 +30,19 @@ const propTypes = {
    * The container to wrap the disclosed content. This should be provided from the application level.
    */
   withDisclosureContainer: PropTypes.func,
+  /**
+   * If set to true, then the focus lock will get enabled.
+   */
+  shouldTrapFocus: PropTypes.bool,
+  /**
+   * If set to true, then the outside click will get enabled.
+   */
+  closeOnOutsideClick: PropTypes.bool,
+};
+
+const defaultProps = {
+  shouldTrapFocus: false,
+  closeOnOutsideClick: false,
 };
 
 const heightFromSize = {
@@ -68,7 +81,7 @@ class ModalManager extends React.Component {
 
   renderModal(manager) {
     const {
-      children, disclosureAccessory, withDisclosureContainer, ...customProps
+      children, disclosureAccessory, withDisclosureContainer, shouldTrapFocus, closeOnOutsideClick, ...customProps
     } = this.props;
     const theme = this.context;
 
@@ -103,9 +116,10 @@ class ModalManager extends React.Component {
             manager.closeDisclosure();
           }}
           closeOnEsc
-          closeOnOutsideClick={false}
-          ariaLabel={headerDataForPresentedComponent?.title || 'Modal'}
+          closeOnOutsideClick={closeOnOutsideClick}
+          ariaLabel={(headerDataForPresentedComponent) ? headerDataForPresentedComponent.title : customProps['aria-label'] || 'Modal'}
           setModalFocusElementRef={this.setModalFocusElementRef}
+          shouldTrapFocus={shouldTrapFocus}
         >
           <ContentContainer
             fill
@@ -148,6 +162,7 @@ class ModalManager extends React.Component {
 }
 
 ModalManager.propTypes = propTypes;
+ModalManager.defaultProps = defaultProps;
 ModalManager.contextType = ThemeContext;
 
 export default ModalManager;