+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Added
+  * Added shouldTrapFocus to be set to 'true'.
+* Changed
+  * Updated the `terra-date-time-picker` example for field label.
+
 ## 1.64.0 - (February 7, 2024)
 
 * Changed